code/AssetManagementAnomalyDetection/requirements.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.0.5
Flask-CORS>=4.0.0
pymysql>=1.1.0
pyodbc>=4.0.34
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.4.0
joblib>=1.3.2
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
